fixed compilation warnings due to int64 being cast into a int

+++ include/zmq.h
@@ -215,7 +215,7 @@ ZMQ_EXPORT int zmq_msg_copy (zmq_msg_t *dest, zmq_msg_t *src);
 ZMQ_EXPORT void *zmq_msg_data (zmq_msg_t *msg);
 ZMQ_EXPORT size_t zmq_msg_size (zmq_msg_t *msg);
 ZMQ_EXPORT int zmq_msg_more (zmq_msg_t *msg);
-ZMQ_EXPORT int zmq_msg_get (zmq_msg_t *msg, int option);
+ZMQ_EXPORT int64_t zmq_msg_get (zmq_msg_t *msg, int option);
 ZMQ_EXPORT int zmq_msg_set (zmq_msg_t *msg, int option, int optval);
